requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
langgraph==0.2.28
llama-index==0.11.14
llama-index-vector-stores-qdrant==0.3.1
qdrant-client==1.11.3
cohere==5.9.2
openai==1.47.0
tiktoken==0.7.0
celery==5.4.0
aiokafka==0.11.0
redis==5.0.8
motor==3.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.2
boto3==1.35.20
botocore==1.35.20
mlflow==2.16.2
wandb==0.17.8
transformers==4.44.2
peft==0.12.0
bitsandbytes==0.43.3
datasets==3.0.1
evaluate==0.4.3
ragas==0.1.21
tree-sitter==0.22.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.4
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
ruff==0.6.8
black==24.8.0
mypy==1.11.2
pre-commit==3.8.0
prometheus-fastapi-instrumentator==7.0.0
sentry-sdk==2.14.0
python-dotenv==1.0.1
structlog==24.4.0
rich==13.8.1
numpy==2.1.1
testcontainers==4.8.1
trl==0.10.1
